Player respawn on death+reloading world correctly

+++ src/Plattform2/Rygar2Engine/src/GameObjectManager.cpp
@@ -99,7 +99,10 @@ void GameObjectManager::RemoveAllGameObjects()
 	AddAndRemoveObjects();
 	for (GameObject* gameObjectToDelete : myGameObjects)
 	{
-		DeleteGameObject(gameObjectToDelete);
+		if (gameObjectToDelete->HasProperty(PropertyKey::eKeepOnLevelReset) == false)
+		{
+			DeleteGameObject(gameObjectToDelete);
+		}
 	}
 	AddAndRemoveObjects();
 }
